Prototyped the filter fps_precisely
Started from copying the filter fps. Remove all options except fps.

This filter outputs frames only when the pts of the input frame is the
nearest to the output pts, no matter it is before or after.
While the old one doesn't count the frame after the output pts.

Instead of converting input pts to base on output timebase, it computes the
lcm of the input output timescale, and computes frame timestamp base on that.